✨ Feat: download & 환경변수 추가
