+++ backend/.env.example
@@ -16,3 +16,7 @@ JWT_REFRESH_EXPIRES=2days
 ADMIN_ID=admin
 ADMIN_PW=1234
 
+## FTP서버의 설정 HOST/USER/PASSWORD
+FTP_HOST=localhost
+FTP_USER=ftpuser
+FTP_PASSWORD=1234
